test: move filesystem contracts out of unit suite
