+++ tests/protocol/test_version_contract.py
@@ -7,7 +7,8 @@ from version import __version__
 
 
 def test_every_static_capability_manifest_matches_project_version() -> None:
-    manifests = json.loads(Path("tools/tool_manifests.json").read_text(encoding="utf-8"))
+    manifest_path = Path(__file__).resolve().parents[2] / "tools" / "tool_manifests.json"
+    manifests = json.loads(manifest_path.read_text(encoding="utf-8"))
     stale = {
         name: manifest.get("extensions", {}).get("tool_version")
         for name, manifest in manifests.items()
